@@ -23,7 +23,7 @@
     "babel-plugin-transform-object-rest-spread": "^6.23.0",
     "babel-plugin-transform-runtime": "^6.15.0",
     "babel-preset-env": "^1.2.1",
-    "eslint": "^4.14.0",
+    "eslint": "^5.2.0",
     "eslint-config-mindhive": "^2.0.1"
   },
   "peerDependencies": {
